better comments
