+++ packages/kbn-securitysolution-grouping/src/containers/query/helpers.test.ts
