+++ packages/kbn-securitysolution-grouping/src/containers/query/helpers.ts
@@ -6,4 +6,75 @@
  * Side Public License, v 1.
  */
 
+import { Filter, FILTERS } from '@kbn/es-query';
 export const getEmptyValue = () => '—';
+
+export const createGroupFilter = (selectedGroup: string, values?: string[] | null): Filter[] =>
+  values != null && values.length > 0
+    ? values.reduce(
+        (acc: Filter[], query) => [
+          ...acc,
+          {
+            meta: {
+              alias: null,
+              disabled: false,
+              key: selectedGroup,
+              negate: false,
+              params: {
+                query,
+              },
+              type: 'phrase',
+            },
+            query: {
+              match_phrase: {
+                [selectedGroup]: {
+                  query,
+                },
+              },
+            },
+          },
+        ],
+        [
+          {
+            meta: {
+              alias: null,
+              disabled: false,
+              type: FILTERS.CUSTOM,
+              negate: false,
+              key: selectedGroup,
+            },
+            query: {
+              script: {
+                script: {
+                  // this will give us an exact match for events with multiple values on the group field
+                  // for example, when values === ['a'], we match events with ['a'], but not ['a', 'b', 'c']
+                  source: "doc[params['field']].size()==params['size']",
+                  params: {
+                    field: selectedGroup,
+                    size: values.length,
+                  },
+                },
+              },
+            },
+          },
+        ]
+      )
+    : [];
+
+export const getNullGroupFilter = (selectedGroup: string): Filter[] => [
+  {
+    meta: {
+      disabled: false,
+      negate: true,
+      alias: null,
+      key: selectedGroup,
+      value: 'exists',
+      type: 'exists',
+    },
+    query: {
+      exists: {
+        field: selectedGroup,
+      },
+    },
+  },
+];
